// src/app/generated/cite.api/model/scoringModelPermission.ts
/**
 * Cite API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ScoringModelPermission = 'ViewScoringModel' | 'EditScoringModel' | 'ManageScoringModel';

export const ScoringModelPermission = {
    ViewScoringModel: 'ViewScoringModel' as ScoringModelPermission,
    EditScoringModel: 'EditScoringModel' as ScoringModelPermission,
    ManageScoringModel: 'ManageScoringModel' as ScoringModelPermission
};